+++ app/lforms.tpl.js
@@ -52,10 +52,10 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "  <!--combo style (default is 'combo')-->\n" +
     "  <div ng-switch-default class=\"lf-combo-answer\">\n" +
     "    <input name=\"{{item.question +'_'+ $id}}\" type=\"text\"\n" +
-    "           ng-required=\"item._answerRequired\"\n" +
     "           ng-model=\"item.value\" autocomplete-lhc=\"item._autocompOptions\"\n" +
     "           ng-readonly=\"item._readOnly\" placeholder=\"{{item._toolTip}}\"\n" +
-    "           id=\"{{item._elementId}}\">\n" +
+    "           id=\"{{item._elementId}}\"\n" +
+    "           ng-focus=\"setActiveRow(item)\" ng-blur=\"activeRowOnBlur(item)\">\n" +
     "  </div>\n" +
     "</div>\n"
   );
@@ -175,58 +175,51 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "                  ng-switch on=\"cell.dataType\">\n" +
     "                <ng-form name=\"innerForm2\">\n" +
     "                  <div class=\"lf-form-item-data tooltipContainer\">\n" +
-    "                    <span class=\"tooltipContent\">\n" +
-    "                      <div class=\"errorMsg errorRequired\">\"{{ cell.question }}\" requires a value.</div>\n" +
-    "                      <div class=\"errorMsg errorPattern\">\"{{ cell.question }}\" requires a text pattern.</div>\n" +
-    "                      <div class=\"errorMsg errorMax\">\"{{ cell.question }}\" requires a maximum value of ??.</div>\n" +
-    "                      <div class=\"errorMsg errorMin\">\"{{ cell.question }}\" requires a minimum value of ??.</div>\n" +
-    "                      <div class=\"errorMsg errorMaxlength\">\"{{ cell.question }}\" requires a maximum length of ??.</div>\n" +
-    "                      <div class=\"errorMsg errorMinlength\">\"{{ cell.question }}\" requires a minimum length of ??.</div>\n" +
-    "                      <div class=\"errorMsg errorURL\">\"{{ cell.question }}\" requires a url.</div>\n" +
-    "                      <div class=\"errorMsg errorEmail\">\"{{ cell.question }}\" requires an email.</div>\n" +
-    "                      <div class=\"errorMsg errorNumber\">\"{{ cell.question }}\" requires a numeric value.</div>\n" +
-    "                      <div class=\"errorMsg errorREAL\">\"{{ cell.question }}\" requires a decimal value.</div>\n" +
-    "                      <div class=\"errorMsg errorINT\">\"{{ cell.question }}\" requires a integer value.</div>\n" +
-    "                      <div class=\"errorMsg errorTM\">\"{{ cell.question }}\" requires a time value.</div>\n" +
-    "                      <div class=\"errorMsg errorDT\">\"{{ cell.question }}\" requires a date value.</div>\n" +
-    "                    </span>  <!-- validation error messages -->\n" +
+    "                    <div class=\"tooltipContent\" lf-validate=\"cell\" ng-model=\"cell.value\" ng-if=\"cell._hasValidation\"></div>\n" +
     "                    <span ng-switch-when=\"\" > </span>\n" +
     "                    <input ng-switch-when=\"CNE\" name=\"{{cell.question + '_' + $id}}\" type=\"text\"\n" +
-    "                           ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                           ng-model=\"cell.value\"\n" +
     "                           autocomplete-lhc=\"cell._autocompOptions\"\n" +
     "                           ng-readonly=\"cell._readOnly\" placeholder=\"{{cell._toolTip}}\"\n" +
     "                           id=\"{{cell._elementId}}\"\n" +
-    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
+    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                           ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
     "                    <input ng-switch-when=\"CWE\" name=\"{{cell.question + '_' + $id}}\" type=\"text\"\n" +
-    "                           ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                           ng-model=\"cell.value\"\n" +
     "                           autocomplete-lhc=\"cell._autocompOptions\"\n" +
     "                           ng-readonly=\"cell._readOnly\" placeholder=\"{{cell._toolTip}}\"\n" +
     "                           id=\"{{cell._elementId}}\"\n" +
-    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
-    "                    <input ng-switch-when=\"REAL\" name=\"{{cell.question}}\" type=\"number\"\n" +
-    "                           ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                           ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
+    "                    <input ng-switch-when=\"REAL\" name=\"{{cell.question}}\" type=\"text\"\n" +
+    "                           ng-model=\"cell.value\"\n" +
     "                           placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                           id=\"{{cell._elementId}}\"\n" +
-    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
-    "                    <input ng-switch-when=\"INT\" name=\"{{cell.question}}\" type=\"number\"\n" +
-    "                           ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                           ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
+    "                    <input ng-switch-when=\"INT\" name=\"{{cell.question}}\" type=\"text\"\n" +
+    "                           ng-model=\"cell.value\"\n" +
     "                           placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                           id=\"{{cell._elementId}}\"\n" +
-    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
+    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                           ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
     "                    <input ng-switch-when=\"DT\" name=\"{{cell.question}}\" type=\"text\"\n" +
-    "                           ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                           ng-model=\"cell.value\"\n" +
     "                           lf-date=\"dateOptions\" placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                           id=\"{{cell._elementId}}\"\n" +
-    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
-    "                    <textarea ng-switch-when=\"TX\" name=\"{{cell.question}}\" ng-required=\"cell._answerRequired\"\n" +
+    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                           ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
+    "                    <textarea ng-switch-when=\"TX\" name=\"{{cell.question}}\"\n" +
     "                              ng-model=\"cell.value\" placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                              id=\"{{cell._elementId}}\"\n" +
     "                              aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
-    "                              ng-keyup=\"autoExpand($event)\" rows=\"1\"></textarea>\n" +
-    "                    <input ng-switch-default name=\"{{cell.question}}\" ng-required=\"cell._answerRequired\" type=\"text\"\n" +
+    "                              ng-keyup=\"autoExpand($event)\" rows=\"1\"\n" +
+    "                              ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\"></textarea>\n" +
+    "                    <input ng-switch-default name=\"{{cell.question}}\" type=\"text\"\n" +
     "                           ng-model=\"cell.value\" placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                           id=\"{{cell._elementId}}\"\n" +
-    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
+    "                           aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                           ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
     "                  </div>\n" +
     "                </ng-form>\n" +
     "              </td>\n" +
@@ -281,58 +274,51 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "                ng-switch on=\"cell.dataType\">\n" +
     "              <ng-form name=\"innerForm2\">\n" +
     "                <div class=\"lf-form-item-data tooltipContainer\">\n" +
-    "                  <span class=\"tooltipContent\">\n" +
-    "                    <div class=\"errorMsg errorRequired\">\"{{ cell.question }}\" requires a value.</div>\n" +
-    "                    <div class=\"errorMsg errorPattern\">\"{{ cell.question }}\" requires a text pattern.</div>\n" +
-    "                    <div class=\"errorMsg errorMax\">\"{{ cell.question }}\" requires a maximum value of ??.</div>\n" +
-    "                    <div class=\"errorMsg errorMin\">\"{{ cell.question }}\" requires a minimum value of ??.</div>\n" +
-    "                    <div class=\"errorMsg errorMaxlength\">\"{{ cell.question }}\" requires a maximum length of ??.</div>\n" +
-    "                    <div class=\"errorMsg errorMinlength\">\"{{ cell.question }}\" requires a minimum length of ??.</div>\n" +
-    "                    <div class=\"errorMsg errorURL\">\"{{ cell.question }}\" requires a url.</div>\n" +
-    "                    <div class=\"errorMsg errorEmail\">\"{{ cell.question }}\" requires an email.</div>\n" +
-    "                    <div class=\"errorMsg errorNumber\">\"{{ cell.question }}\" requires a numeric value.</div>\n" +
-    "                    <div class=\"errorMsg errorREAL\">\"{{ cell.question }}\" requires a decimal value.</div>\n" +
-    "                    <div class=\"errorMsg errorINT\">\"{{ cell.question }}\" requires a integer value.</div>\n" +
-    "                    <div class=\"errorMsg errorTM\">\"{{ cell.question }}\" requires a time value.</div>\n" +
-    "                    <div class=\"errorMsg errorDT\">\"{{ cell.question }}\" requires a date value.</div>\n" +
-    "                  </span>  <!-- validation error messages -->\n" +
+    "                  <div class=\"tooltipContent\" lf-validate=\"cell\" ng-model=\"cell.value\" ng-if=\"cell._hasValidation\"></div>\n" +
     "                  <span ng-switch-when=\"\" > </span>\n" +
     "                  <input ng-switch-when=\"CNE\" name=\"{{cell.question + '_' + $id}}\" type=\"text\"\n" +
-    "                         ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                         ng-model=\"cell.value\"\n" +
     "                         autocomplete-lhc=\"cell._autocompOptions\"\n" +
     "                         ng-readonly=\"cell._readOnly\" placeholder=\"{{cell._toolTip}}\"\n" +
     "                         id=\"{{cell._elementId}}\"\n" +
-    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
+    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                         ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
     "                  <input ng-switch-when=\"CWE\" name=\"{{cell.question + '_' + $id}}\" type=\"text\"\n" +
-    "                         ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                         ng-model=\"cell.value\"\n" +
     "                         autocomplete-lhc=\"cell._autocompOptions\"\n" +
     "                         ng-readonly=\"cell._readOnly\" placeholder=\"{{cell._toolTip}}\"\n" +
     "                         id=\"{{cell._elementId}}\"\n" +
-    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
-    "                  <input ng-switch-when=\"REAL\" name=\"{{cell.question}}\" type=\"number\"\n" +
-    "                         ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                         ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
+    "                  <input ng-switch-when=\"REAL\" name=\"{{cell.question}}\" type=\"text\"\n" +
+    "                         ng-model=\"cell.value\"\n" +
     "                         placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                         id=\"{{cell._elementId}}\"\n" +
-    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
-    "                  <input ng-switch-when=\"INT\" name=\"{{cell.question}}\" type=\"number\"\n" +
-    "                         ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                         ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
+    "                  <input ng-switch-when=\"INT\" name=\"{{cell.question}}\" type=\"text\"\n" +
+    "                         ng-model=\"cell.value\"\n" +
     "                         placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                         id=\"{{cell._elementId}}\"\n" +
-    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
+    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                         ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
     "                  <input ng-switch-when=\"DT\" name=\"{{cell.question}}\" type=\"text\"\n" +
-    "                         ng-required=\"cell._answerRequired\" ng-model=\"cell.value\"\n" +
+    "                         ng-model=\"cell.value\"\n" +
     "                         lf-date=\"dateOptions\" placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                         id=\"{{cell._elementId}}\"\n" +
-    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
-    "                  <textarea ng-switch-when=\"TX\" name=\"{{cell.question}}\" ng-required=\"cell._answerRequired\"\n" +
+    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                         ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
+    "                  <textarea ng-switch-when=\"TX\" name=\"{{cell.question}}\"\n" +
     "                            ng-model=\"cell.value\" placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                            id=\"{{cell._elementId}}\"\n" +
     "                            aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
-    "                            ng-keyup=\"autoExpand($event)\" rows=\"1\"></textarea>\n" +
-    "                  <input ng-switch-default name=\"{{cell.question}}\" ng-required=\"cell._answerRequired\" type=\"text\"\n" +
+    "                            ng-keyup=\"autoExpand($event)\" rows=\"1\"\n" +
+    "                            ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\"></textarea>\n" +
+    "                  <input ng-switch-default name=\"{{cell.question}}\" type=\"text\"\n" +
     "                         ng-model=\"cell.value\" placeholder=\"{{cell._toolTip}}\" ng-readonly=\"cell._readOnly\"\n" +
     "                         id=\"{{cell._elementId}}\"\n" +
-    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\">\n" +
+    "                         aria-labelledby=\"{{lfData._horizontalTableInfo[item._horizontalTableId].columnHeaders[$index].id}}\"\n" +
+    "                         ng-focus=\"setActiveRow(cell)\" ng-blur=\"activeRowOnBlur(cell)\">\n" +
     "                </div>\n" +
     "              </ng-form>\n" +
     "            </td>\n" +
@@ -352,7 +338,7 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "              ng-click=\"addOneRepeatingItem(item, true)\"\n" +
     "              ng-blur=\"hideUnusedItemWarning(item)\"\n" +
     "              uib-popover='Please enter info in the blank \"{{ item.question }}\".'\n" +
-    "              popover-placement=\"top-left\" popover-title=\"Warning\"\n" +
+    "              popover-placement=\"top-left\"\n" +
     "              popover-trigger=\"none\"\n" +
     "              popover-is-open=\"item._showUnusedItemWarning\">\n" +
     "        Add another '{{item.question}}'\n" +
@@ -485,8 +471,7 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "  <!--input field-->\n" +
     "  <ng-form name=\"innerForm2\" class=\"hasTooltip\">\n" +
     "    <div class=\"lf-form-item-data tooltipContainer\" ng-switch on=\"item.dataType\">\n" +
-    "      <span class=\"tooltipContent\" ng-include=\"'validation.html'\"></span>  <!-- validation error messages -->\n" +
-    "      <!--<span class=\"tooltipContent\" lf-validation></span>-->\n" +
+    "      <div class=\"tooltipContent\" lf-validate=\"item\" ng-model=\"item.value\" ng-if=\"item._hasValidation\"></div>\n" +
     "      <span ng-switch-when=\"SECTION\" id=\"{{item._elementId}}\"> </span>\n" +
     "      <span ng-switch-when=\"TTILE\" id=\"{{item._elementId}}\"> </span>\n" +
     "      <div ng-switch-when=\"CNE\">\n" +
@@ -495,24 +480,32 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "      <div ng-switch-when=\"CWE\">\n" +
     "        <lf-answers item=\"item\"></lf-answers>\n" +
     "      </div>\n" +
-    "      <input ng-switch-when=\"REAL\" name=\"{{item.question}}\" type=\"number\"\n" +
-    "             ng-required=\"item._answerRequired\"\n" +
+    "      <input ng-switch-when=\"REAL\" name=\"{{item.question}}\" type=\"text\"\n" +
     "             ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\"\n" +
-    "             ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\">\n" +
-    "      <input ng-switch-when=\"INT\" name=\"{{item.question}}\" type=\"number\"\n" +
-    "             ng-required=\"item._answerRequired\"\n" +
+    "             ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\"\n" +
+    "             ng-focus=\"setActiveRow(item)\"\n" +
+    "             ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "      <input ng-switch-when=\"INT\" name=\"{{item.question}}\" type=\"text\"\n" +
     "             ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\"\n" +
-    "             ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\">\n" +
-    "      <input ng-switch-when=\"DT\" name=\"{{item.question}}\" ng-required=\"item._answerRequired\" type=\"text\"\n" +
+    "             ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\"\n" +
+    "             ng-focus=\"setActiveRow(item)\"\n" +
+    "             ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "      <input ng-switch-when=\"DT\" name=\"{{item.question}}\" type=\"text\"\n" +
     "             ng-model=\"item.value\" lf-date=\"dateOptions\" placeholder=\"{{item._toolTip}}\"\n" +
-    "             ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\">\n" +
-    "      <textarea ng-switch-when=\"TX\" name=\"{{item.question}}\" ng-required=\"item._answerRequired\"\n" +
+    "             ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\"\n" +
+    "             ng-focus=\"setActiveRow(item)\"\n" +
+    "             ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "      <textarea ng-switch-when=\"TX\" name=\"{{item.question}}\"\n" +
     "                ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\" ng-readonly=\"item._readOnly\"\n" +
-    "                id=\"{{item._elementId}}\" ng-keyup=\"autoExpand($event)\" ng-blur=\"autoExpand($event)\" rows=\"1\">\n" +
+    "                id=\"{{item._elementId}}\" ng-keyup=\"autoExpand($event)\" ng-blur=\"autoExpand($event)\" rows=\"1\"\n" +
+    "                ng-focus=\"setActiveRow(item)\"\n" +
+    "                ng-blur=\"activeRowOnBlur(item)\">\n" +
     "      </textarea>\n" +
-    "      <input ng-switch-default name=\"{{item.question}}\" ng-required=\"item._answerRequired\" type=\"text\"\n" +
+    "      <input ng-switch-default name=\"{{item.question}}\" type=\"text\"\n" +
     "             ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\" ng-readonly=\"item._readOnly\"\n" +
-    "             id=\"{{item._elementId}}\">\n" +
+    "             id=\"{{item._elementId}}\"\n" +
+    "             ng-focus=\"setActiveRow(item)\"\n" +
+    "             ng-blur=\"activeRowOnBlur(item)\">\n" +
     "    </div>\n" +
     "  </ng-form>\n" +
     "  <!--unit-->\n" +
@@ -524,13 +517,14 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "  <!--extra question-->\n" +
     "  <div ng-if=\"needExtra(item) && targetShown(item)\" ng-click=\"setActiveRow(item)\"\n" +
     "       class=\"extra-row {{getRowClass(item)}} {{getSkipLogicClass(item)}} {{getActiveRowClass(item)}}\">\n" +
-    "    <input ng-model=\"item.valueOther\" placeholder=\"Please specify\" ng-readonly=\"item._readOnly\">\n" +
+    "    <input ng-model=\"item.valueOther\" placeholder=\"Please specify\" ng-readonly=\"item._readOnly\"\n" +
+    "           ng-focus=\"setActiveRow(item)\">\n" +
     "  </div>\n" +
     "  <!--sub sections, check each item's layout -->\n" +
     "  <div ng-if=\"item.items\" class=\"section\">\n" +
     "    <div ng-repeat=\"item in item.items\" ng-if=\"targetShown(item)\"\n" +
-    "         class=\"data-row has-ng-animate {{getRowClass(item)}} {{getSkipLogicClass(item)}} {{getActiveRowClass(item)}}\n" +
-    "             ng-switch on item.displayControl.questionLayout\">\n" +
+    "         class=\"data-row has-ng-animate {{getRowClass(item)}} {{getSkipLogicClass(item)}} {{getActiveRowClass(item)}}\"\n" +
+    "         ng-click=\"setActiveRow(item)\">\n" +
     "      <div ng-if=\"item.header\" ng-switch on=\"item.displayControl.questionLayout\">\n" +
     "        <div ng-switch-when=\"horizontal\">\n" +
     "          <lf-section-horizontal></lf-section-horizontal>\n" +
@@ -561,7 +555,7 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "              ng-click=\"addOneRepeatingItem(item)\"\n" +
     "              ng-blur=\"hideUnusedItemWarning(item)\"\n" +
     "              uib-popover='Please enter info in the blank \"{{ item.question }}\".'\n" +
-    "              popover-placement=\"top-left\" popover-title=\"Warning\"\n" +
+    "              popover-placement=\"top-left\"\n" +
     "              popover-trigger=\"none\"\n" +
     "              popover-is-open=\"item._showUnusedItemWarning\">\n" +
     "        Add another '{{item.question}}'\n" +
@@ -608,8 +602,8 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "          <!-- check each top level item's layout -->\n" +
     "          <div ng-if=\"lfData.items\">\n" +
     "            <div ng-repeat=\"item in lfData.items\" ng-if=\"targetShown(item)\"\n" +
-    "                 class=\"data-row has-ng-animate {{getRowClass(item)}} {{getSkipLogicClass(item)}} {{getActiveRowClass(item)}}\n" +
-    "             ng-switch on item.displayControl.questionLayout\">\n" +
+    "                 class=\"data-row has-ng-animate {{getRowClass(item)}} {{getSkipLogicClass(item)}} {{getActiveRowClass(item)}}\"\n" +
+    "                 ng-click=\"setActiveRow(item)\">\n" +
     "              <div ng-if=\"item.header\" ng-switch on=\"item.displayControl.questionLayout\">\n" +
     "                <div ng-switch-when=\"horizontal\">\n" +
     "                  <lf-section-horizontal></lf-section-horizontal>\n" +
@@ -669,26 +663,27 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "        </tr>\n" +
     "        </thead>\n" +
     "        <tbody>\n" +
-    "        <tr class=\"lf-form-table-row\" ng-if=\"lfData.templateOptions.obrHeader\">\n" +
+    "        <tr class=\"lf-form-table-row\" ng-if=\"lfData.templateOptions.obrHeader\" ng-click=\"setActiveRow(null)\">\n" +
     "          <td class=\"rowEditText hasTooltip\" ng-repeat=\"item in lfData.templateOptions.obrItems\"\n" +
     "              ng-switch on=\"item.dataType\">\n" +
     "            <ng-form name=\"innerForm\">\n" +
     "              <div class=\"lf-form-item-data tooltipContainer\">\n" +
-    "                <span class=\"tooltipContent\" ng-include=\"'validation.html'\"></span>  <!-- validation error messages -->\n" +
+    "                <div class=\"tooltipContent\" lf-validate=\"item\" ng-model=\"item.value\"></div>\n" +
     "                <input ng-switch-when=\"CWE\" name=\"{{item.question}}\" type=\"text\"\n" +
-    "                       ng-required=\"item._answerRequired\" placeholder=\"Select or type a value\"\n" +
+    "                       placeholder=\"Select or type a value\"\n" +
     "                       ng-model=\"item.value\"\n" +
     "                       autocomplete-lhc=\"item._autocompOptions\"\n" +
-    "                       id=\"{{item.questionCode}}\">\n" +
+    "                       id=\"{{item.questionCode}}\"\n" +
+    "                       ng-blur=\"activeRowOnBlur(item)\">\n" +
     "                <input ng-switch-when=\"DT\" name=\"{{item.question}}\" type=\"text\"\n" +
-    "                       ng-required=\"item._answerRequired\"\n" +
     "                       ng-model=\"item.value\" lf-date=\"dateOptions\"\n" +
     "                       placeholder=\"MM/DD/YYYY\"\n" +
-    "                       id=\"{{item.questionCode}}\">\n" +
+    "                       id=\"{{item.questionCode}}\"\n" +
+    "                       ng-blur=\"activeRowOnBlur(item)\">\n" +
     "                <input ng-switch-default name=\"{{item.question}}\" type=\"text\"\n" +
-    "                       ng-required=\"item._answerRequired\"\n" +
     "                       ng-model=\"item.value\" placeholder=\"Type a value\"\n" +
-    "                       id=\"{{item.questionCode}}\">\n" +
+    "                       id=\"{{item.questionCode}}\"\n" +
+    "                       ng-blur=\"activeRowOnBlur(item)\">\n" +
     "              </div>\n" +
     "            </ng-form>\n" +
     "          </td>\n" +
@@ -750,36 +745,40 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "                  <td ng-switch on=\"item.dataType\" class=\"hasTooltip\">\n" +
     "                    <ng-form name=\"innerForm2\">\n" +
     "                      <div class=\"lf-form-item-data tooltipContainer\">\n" +
-    "                        <span class=\"tooltipContent\" ng-include=\"'validation.html'\"></span>  <!-- validation error messages -->\n" +
+    "                        <div class=\"tooltipContent\" lf-validate=\"item\" ng-model=\"item.value\" ng-if=\"item._hasValidation\"></div>\n" +
     "                        <span ng-switch-when=\"SECTION\" id=\"{{item._elementId}}\"> </span>\n" +
-    "                        <input ng-switch-when=\"CNE\" name=\"{{item.question +'_'+ $id}}\"\n" +
-    "                               ng-required=\"item._answerRequired\" type=\"text\"\n" +
+    "                        <input ng-switch-when=\"CNE\" name=\"{{item.question +'_'+ $id}}\" type=\"text\"\n" +
     "                               ng-model=\"item.value\" autocomplete-lhc=\"item._autocompOptions\"\n" +
     "                               ng-readonly=\"item._readOnly\" placeholder=\"{{item._toolTip}}\"\n" +
-    "                               id=\"{{item._elementId}}\">\n" +
-    "                        <input ng-switch-when=\"CWE\" name=\"{{item.question +'_'+ $id}}\"\n" +
-    "                               ng-required=\"item._answerRequired\" type=\"text\"\n" +
+    "                               id=\"{{item._elementId}}\" ng-focus=\"setActiveRow(item)\"\n" +
+    "                               ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "                        <input ng-switch-when=\"CWE\" name=\"{{item.question +'_'+ $id}}\" type=\"text\"\n" +
     "                               ng-model=\"item.value\" autocomplete-lhc=\"item._autocompOptions\"\n" +
     "                               ng-readonly=\"item._readOnly\" placeholder=\"{{item._toolTip}}\"\n" +
-    "                               id=\"{{item._elementId}}\">\n" +
-    "                        <input ng-switch-when=\"REAL\" name=\"{{item.question}}\" type=\"number\"\n" +
-    "                               ng-required=\"item._answerRequired\"\n" +
+    "                               id=\"{{item._elementId}}\" ng-focus=\"setActiveRow(item)\"\n" +
+    "                               ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "                        <input ng-switch-when=\"REAL\" name=\"{{item.question}}\" type=\"text\"\n" +
     "                               ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\"\n" +
-    "                               ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\">\n" +
-    "                        <input ng-switch-when=\"INT\" name=\"{{item.question}}\" type=\"number\"\n" +
-    "                               ng-required=\"item._answerRequired\"\n" +
+    "                               ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\" ng-focus=\"setActiveRow(item)\"\n" +
+    "                               ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "                        <input ng-switch-when=\"INT\" name=\"{{item.question}}\" type=\"text\"\n" +
     "                               ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\"\n" +
-    "                               ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\">\n" +
-    "                        <input ng-switch-when=\"DT\" name=\"{{item.question}}\" ng-required=\"item._answerRequired\" type=\"text\"\n" +
+    "                               ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\" ng-focus=\"setActiveRow(item)\"\n" +
+    "                               ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "                        <input ng-switch-when=\"DT\" name=\"{{item.question}}\" type=\"text\"\n" +
     "                               ng-model=\"item.value\" lf-date=\"dateOptions\" placeholder=\"{{item._toolTip}}\"\n" +
-    "                               ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\">\n" +
-    "                        <textarea ng-switch-when=\"TX\" name=\"{{item.question}}\" ng-required=\"item._answerRequired\"\n" +
+    "                               ng-readonly=\"item._readOnly\" id=\"{{item._elementId}}\" ng-focus=\"setActiveRow(item)\"\n" +
+    "                               ng-blur=\"activeRowOnBlur(item)\">\n" +
+    "                        <textarea ng-switch-when=\"TX\" name=\"{{item.question}}\"\n" +
     "                                  ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\" ng-readonly=\"item._readOnly\"\n" +
-    "                                  id=\"{{item._elementId}}\" ng-keyup=\"autoExpand($event)\" ng-blur=\"autoExpand($event)\" rows=\"1\">\n" +
+    "                                  id=\"{{item._elementId}}\" ng-keyup=\"autoExpand($event)\" ng-blur=\"autoExpand($event)\" rows=\"1\"\n" +
+    "                                  ng-focus=\"setActiveRow(item)\"\n" +
+    "                                  ng-blur=\"activeRowOnBlur(item)\">\n" +
     "                        </textarea>\n" +
-    "                        <input ng-switch-default name=\"{{item.question}}\" ng-required=\"item._answerRequired\" type=\"text\"\n" +
+    "                        <input ng-switch-default name=\"{{item.question}}\" type=\"text\"\n" +
     "                               ng-model=\"item.value\" placeholder=\"{{item._toolTip}}\" ng-readonly=\"item._readOnly\"\n" +
-    "                               id=\"{{item._elementId}}\">\n" +
+    "                               id=\"{{item._elementId}}\" ng-focus=\"setActiveRow(item)\"\n" +
+    "                               ng-blur=\"activeRowOnBlur(item)\">\n" +
     "                      </div>\n" +
     "                    </ng-form>\n" +
     "                  </td>\n" +
@@ -846,7 +845,7 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "                  </td>\n" +
     "                  <td class=\"button-col\"></td>\n" +
     "                  <td colspan=\"5\" class=\"extra-field\">\n" +
-    "                    <input ng-model=\"item.valueOther\" placeholder=\"Please specify\" ng-readonly=\"item._readOnly\" type=\"text\">\n" +
+    "                    <input ng-model=\"item.valueOther\" placeholder=\"Please specify\" ng-readonly=\"item._readOnly\" type=\"text\" ng-focus=\"setActiveRow(item)\">\n" +
     "                  </td>\n" +
     "                </tr>\n" +
     "                <!--a button row at the end of each repeating section-->\n" +
@@ -865,7 +864,7 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "                                    ng-click=\"addOneRepeatingItem(repeatingItem)\"\n" +
     "                                    ng-blur=\"hideUnusedItemWarning(repeatingItem)\"\n" +
     "                                    uib-popover='Please enter info in the blank \"{{ repeatingItem.question }}\".'\n" +
-    "                                    popover-placement=\"top-left\" popover-title=\"Warning\"\n" +
+    "                                    popover-placement=\"top-left\"\n" +
     "                                    popover-trigger=\"none\"\n" +
     "                                    popover-is-open=\"repeatingItem._showUnusedItemWarning\">\n" +
     "                              Add another '{{repeatingItem.question}}'\n" +
@@ -887,24 +886,6 @@ angular.module('lformsWidget').run(['$templateCache', function($templateCache) {
     "  </div>\n" +
     "</form>\n" +
     "<button type=\"button\" ng-if=\"debug\" ng-click=\"onclick()\">Click to debug Panel Controller</button>\n"
-  );
-
-
-  $templateCache.put('validation.html',
-    "<div class=\"errorMsg errorRequired\">\"{{ item.question }}\" requires a value.</div>\n" +
-    "<div class=\"errorMsg errorPattern\">\"{{ item.question }}\" requires a text pattern.</div>\n" +
-    "<div class=\"errorMsg errorMax\">\"{{ item.question }}\" requires a maximum value of ??.</div>\n" +
-    "<div class=\"errorMsg errorMin\">\"{{ item.question }}\" requires a minimum value of ??.</div>\n" +
-    "<div class=\"errorMsg errorMaxlength\">\"{{ item.question }}\" requires a maximum length of ??.</div>\n" +
-    "<div class=\"errorMsg errorMinlength\">\"{{ item.question }}\" requires a minimum length of ??.</div>\n" +
-    "<div class=\"errorMsg errorURL\">\"{{ item.question }}\" requires a url.</div>\n" +
-    "<div class=\"errorMsg errorEmail\">\"{{ item.question }}\" requires an email.</div>\n" +
-    "<div class=\"errorMsg errorNumber\">\"{{ item.question }}\" requires a numeric value.</div>\n" +
-    "<div class=\"errorMsg errorREAL\">\"{{ item.question }}\" requires a decimal value.</div>\n" +
-    "<div class=\"errorMsg errorINT\">\"{{ item.question }}\" requires a integer value.</div>\n" +
-    "<div class=\"errorMsg errorTM\">\"{{ item.question }}\" requires a time value.</div>\n" +
-    "<div class=\"errorMsg errorDT\">\"{{ item.question }}\" requires a date value.</div>\n" +
-    "<div class=\"errorMsg errorInvalidParse\">\"{{ item.question }}\" requires a selection from the list.</div>\n"
   );
 
 }]);