+++ app/scripts/lib/lforms-util.js
@@ -1,0 +1,100 @@
+/**
+ * LForms Utility tools
+ */
+if (typeof LForms === 'undefined')
+  LForms = {};
+
+LForms.Util = {
+
+ /**
+   *  Adds an LForms form to the page.
+   * @param formDataVar The name of a global-scope variable containing the
+   *  form's LForms definition.  The variable should be accessible as a property
+   *  of the window object.
+   * @param formContainer The ID of a DOM element to contain the form, or the
+   *  element itself.  The contents of this element will be replaced by the form.
+   *  This element should be outside the scope of any existing AngularJS app on
+   *  the page.
+   */
+  addFormToPage: function(formDataVar, formContainer) {
+    var formContainer = typeof formContainer === 'string' ?
+      $('#'+formContainer) : $(formContainer);
+    if (!this.pageFormID_)
+      this.pageFormID_ = 0;
+    var appName = 'LFormsApp' + ++this.pageFormID_;
+    var controller = 'LFormsAppController'+ this.pageFormID_;
+    formContainer.html(
+      '<div ng-controller="'+controller+'">'+
+        '<lforms lf-data="myFormData"></lforms>'+
+      '</div>'+
+      '<script>'+
+        'angular.module("'+appName+'", ["lformsWidget"])'+
+        '.controller("'+controller+'", ["$scope", function ($scope) {'+
+        '  $scope.myFormData = new LForms.LFormsData('+formDataVar+');'+
+        '}]);'+
+      '</'+'script>'
+    );
+    // Bootstrap the element if needed
+    // Following http://stackoverflow.com/a/34501500/360782
+    var isInitialized = formContainer.injector();
+    if (!isInitialized)
+      angular.bootstrap(formContainer.children(':first'), [appName]);
+  },
+
+
+  /**
+   * Get user input data from the form, with or without form definition data.
+   * @param element the containing HTML element that includes the LForm's rendered form.
+   * @param noFormDefData optional, to include form definition data, the default is false.
+   * @param noEmptyValue optional, to remove items that have an empty value, the default is false.
+   * @param noHiddenItem optional, to remove items that are hidden by skip logic, the default is false.
+   * @param keepIdPath optional, to keep _idPath field on item
+   * @returns {{itemsData: (*|Array), templateData: (*|Array)}} form data and template data
+   */
+  getUserData: function(element, noFormDefData, noEmptyValue, noHiddenItem, keepIdPath) {
+    var formObj = this._getFormObjectInScope(element);
+    return formObj ? formObj.getUserData(noFormDefData, noEmptyValue, noHiddenItem, keepIdPath) : null;
+  },
+
+
+  /**
+   * Get the complete form definition data, including the user input data from the form.
+   * The returned data could be fed into a LForms widget directly to render the form.
+   * @param element required, the containing HTML element that includes the LForm's rendered form.
+   *        It could be the DOM element or its id.
+   * @param noEmptyValue optional, to remove items that have an empty value, the default is false.
+   * @param noHiddenItem optional, to remove items that are hidden by skip logic, the default is false.
+   * @param keepIdPath optional, to keep _idPath field on item
+   * @returns {{}} Form definition data
+   */
+  getFormData: function(element, noEmptyValue, noHiddenItem, keepIdPath) {
+    var formObj = this._getFormObjectInScope(element);
+    return formObj ? formObj.getFormData(noEmptyValue, noHiddenItem, keepIdPath) : null;
+  },
+
+
+  /**
+   * Find the form object in the scope based the form dom element
+   * @param element element the containing HTML element that includes the LForm's rendered form.
+   * @returns {*}
+   * @private
+   */
+  _getFormObjectInScope: function(element) {
+    // find the scope that has the LForms data
+    var formObj;
+    if (!element) element = jQuery("body");
+
+    // class="lf-form"> is the element that contains rendered form.
+    var lfForms = jQuery(element).find(".lf-form");
+    angular.forEach(lfForms, function(ele, index) {
+      var lfForm = angular.element(ele);
+      if (lfForm.scope() && lfForm.scope().lfData) {
+        formObj = lfForm.scope().lfData;
+        return false; // break the loop
+      }
+    });
+
+    return formObj;
+  }
+};
+